requirements.txt
numpy~=1.23.5
matplotlib~=3.7.1
plotly~=5.14.1
pillow~=9.4.0
tqdm~=4.65.0
setuptools~=65.6.3
pytorch~=2.0.0
